refactor bp editor into multiple files for maintainability
