+++ lib/features/bp/obs/widgets.dart
@@ -1,0 +1,303 @@
+/// Blood pressure observation widgets.
+//
+// Time-stamp: <Thursday 2024-12-19 13:33:06 +1100 Graham Williams>
+//
+/// Copyright (C) 2025, Software Innovation Institute, ANU
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License");
+///
+/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://www.gnu.org/licenses/>.
+///
+/// Authors: Ashley Tang.
+
+library;
+
+import 'package:flutter/material.dart';
+
+import 'package:intl/intl.dart';
+
+import 'package:healthpod/features/bp/editor/state.dart';
+import 'package:healthpod/features/bp/obs/model.dart';
+import 'package:healthpod/utils/parse_numeric_input.dart';
+
+/// Builds the display (read-only) row for a BP observation.
+
+DataRow buildDisplayRow({
+  required BuildContext context,
+  required BPObservation observation,
+  required int index,
+  required VoidCallback onEdit,
+  required VoidCallback onDelete,
+}) {
+  return DataRow(
+    cells: [
+      DataCell(Text(
+          DateFormat('yyyy-MM-dd HH:mm:ss').format(observation.timestamp))),
+      DataCell(Text(parseNumericInput(observation.systolic))),
+      DataCell(Text(parseNumericInput(observation.diastolic))),
+      DataCell(Text(parseNumericInput(observation.heartRate))),
+      DataCell(Text(observation.feeling)),
+      DataCell(
+        Container(
+          constraints: const BoxConstraints(maxWidth: 200),
+          child: Text(
+            observation.notes,
+            overflow: TextOverflow.ellipsis,
+            maxLines: 3,
+          ),
+        ),
+      ),
+      DataCell(
+        Row(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            IconButton(
+              icon: const Icon(Icons.edit),
+              onPressed: onEdit,
+            ),
+            IconButton(
+              icon: const Icon(Icons.delete),
+              onPressed: onDelete,
+            ),
+          ],
+        ),
+      ),
+    ],
+  );
+}
+
+/// Builds the editing row for a BP observation. Pulls [TextEditingController]s
+/// from [editorState], so that typed-in changes are reflected in the state's
+/// [currentEdit].
+
+DataRow buildEditingRow({
+  required BuildContext context,
+  required BPEditorState editorState,
+  required dynamic editorService,
+  required BPObservation observation,
+  required int index,
+  required VoidCallback onCancel,
+  required VoidCallback onSave,
+}) {
+  final currentEdit = editorState.currentEdit ?? observation;
+
+  // Ensure controllers are initialised.
+
+  if (editorState.systolicController == null) {
+    editorState.initialiseControllers(observation);
+  }
+
+  return DataRow(
+    cells: [
+      // Timestamp cell with date/time picker.
+
+      DataCell(
+        InkWell(
+          onTap: () async {
+            final date = await showDatePicker(
+              context: context,
+              initialDate: currentEdit.timestamp,
+              firstDate: DateTime(2000),
+              lastDate: DateTime.now(),
+            );
+
+            if (!context.mounted) return;
+
+            if (date != null) {
+              final time = await showTimePicker(
+                context: context,
+                initialTime: TimeOfDay.fromDateTime(currentEdit.timestamp),
+              );
+              if (time != null && context.mounted) {
+                // (Optional) Show a dialog for milliseconds if needed.
+
+                final TextEditingController msController =
+                    TextEditingController();
+                final milliseconds = await showDialog<int>(
+                  context: context,
+                  builder: (_) => AlertDialog(
+                    title: const Text('Set Milliseconds'),
+                    content: TextField(
+                      controller: msController,
+                      keyboardType: TextInputType.number,
+                      decoration: const InputDecoration(
+                        hintText: 'Enter milliseconds (0-999)',
+                      ),
+                    ),
+                    actions: [
+                      TextButton(
+                        onPressed: () => Navigator.of(context).pop(0),
+                        child: const Text('Skip'),
+                      ),
+                      TextButton(
+                        onPressed: () {
+                          final ms = int.tryParse(msController.text) ?? 0;
+                          Navigator.of(context).pop(ms.clamp(0, 999));
+                        },
+                        child: const Text('Confirm'),
+                      ),
+                    ],
+                  ),
+                );
+
+                final newTimestamp = DateTime(
+                  date.year,
+                  date.month,
+                  date.day,
+                  time.hour,
+                  time.minute,
+                  0,
+                  milliseconds ?? 0,
+                );
+
+                // Prevent duplicate timestamp.
+
+                if (editorState.observations.any(
+                  (r) =>
+                      r.timestamp == newTimestamp &&
+                      editorState.observations.indexOf(r) != index,
+                )) {
+                  if (!context.mounted) return;
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(
+                      content: Text(
+                          'An observation with this timestamp already exists'),
+                    ),
+                  );
+                  return;
+                }
+
+                // Update state's observation.
+
+                editorState.observations[index] =
+                    observation.copyWith(timestamp: newTimestamp);
+              }
+            }
+          },
+          child: Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8.0),
+            child: Text(
+              DateFormat('yyyy-MM-dd HH:mm:ss.SSS')
+                  .format(observation.timestamp),
+              style: const TextStyle(
+                decoration: TextDecoration.underline,
+                color: Colors.blue,
+              ),
+            ),
+          ),
+        ),
+      ),
+
+      // Systolic.
+
+      DataCell(
+        TextField(
+          controller: editorState.systolicController,
+          keyboardType: const TextInputType.numberWithOptions(decimal: true),
+          onChanged: (value) {
+            editorState.currentEdit = currentEdit.copyWith(
+              systolic: value.isEmpty ? 0 : (double.tryParse(value) ?? 0.0),
+            );
+          },
+        ),
+      ),
+
+      // Diastolic.
+
+      DataCell(
+        TextField(
+          controller: editorState.diastolicController,
+          keyboardType: const TextInputType.numberWithOptions(decimal: true),
+          onChanged: (value) {
+            editorState.currentEdit = currentEdit.copyWith(
+              diastolic: value.isEmpty ? 0 : (double.tryParse(value) ?? 0.0),
+            );
+          },
+        ),
+      ),
+
+      // Heart Rate.
+
+      DataCell(
+        TextField(
+          controller: editorState.heartRateController,
+          keyboardType: const TextInputType.numberWithOptions(decimal: true),
+          onChanged: (value) {
+            editorState.currentEdit = currentEdit.copyWith(
+              heartRate: value.isEmpty ? 0 : (double.tryParse(value) ?? 0.0),
+            );
+          },
+        ),
+      ),
+
+      // Feeling dropdown.
+
+      DataCell(
+        DropdownButton<String>(
+          value: currentEdit.feeling.isEmpty ? null : currentEdit.feeling,
+          items: ['Excellent', 'Good', 'Fair', 'Poor']
+              .map((feeling) => DropdownMenuItem(
+                    value: feeling,
+                    child: Text(feeling),
+                  ))
+              .toList(),
+          onChanged: (value) {
+            editorState.currentEdit = currentEdit.copyWith(
+              feeling: value ?? '',
+            );
+          },
+        ),
+      ),
+
+      // Notes cell.
+
+      DataCell(
+        Container(
+          constraints: const BoxConstraints(maxWidth: 200),
+          child: TextField(
+            controller: editorState.notesController,
+            maxLines: null,
+            decoration: const InputDecoration(
+              isDense: true,
+              contentPadding: EdgeInsets.symmetric(vertical: 8.0),
+            ),
+            onChanged: (value) {
+              editorState.currentEdit = currentEdit.copyWith(notes: value);
+            },
+          ),
+        ),
+      ),
+
+      // Actions cell.
+
+      DataCell(
+        Row(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            IconButton(
+              icon: const Icon(Icons.save),
+              onPressed: onSave,
+            ),
+            IconButton(
+              icon: const Icon(Icons.cancel),
+              onPressed: onCancel,
+            ),
+          ],
+        ),
+      ),
+    ],
+  );
+}
